Improve Link support to Arch BuildingPart
BuildingPart is a special plain group that has its own display mode
(presumably for annotation and stuff). This patch adds a 'tailChild'
field to SoFCSwitch which will be traversed as long as its 'whichChild'
is not -1. The plain group coin3D node tree snapshot uses its field to
render the group children along with group's own display mode.