App: add new property SyncGroupVisibility to App::Link
This property is used to control whether to sync children visibilities
when linked to a group. If not enabled (default), the linked group will
have independent vsibility control to all nested children of the group.
This is the old behavior. With this new property, the user can choose
to sync the children visibility, which makes the linked group looks
exactly the same, execpt with a different placement.